requirements.txt
streamlit
pillow
boto3
pymongo
python-dotenv
openai
twilio
reportlab
